# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Run Coaching System — UI↔Engine API
  version: 0.1.0-draft
  description: >
    Target contract for the athlete-facing web + mobile UI. Read/write surface
    over the deterministic coaching engine. The engine is the single source of
    truth for every training decision; this API only exposes the engine's
    objects and accepts the structured inputs it already consumes. See UI spec
    `spec/ui-00-outline.md` §4 and the engine spec `spec/01`–`spec/09`.
servers:
  - url: http://127.0.0.1:8000
    description: Local backend (retsaf-nur runcoach-api default)
  - url: https://api.example.invalid
    description: Placeholder for a hosted deployment (TBD)
...
security:
  - bearerAuth: []
paths:
...
  /sessions:
    post:
      tags: [Sessions]
      operationId: ingestSession
      summary: Ingest a Garmin FIT file into the canonical record (manual-upload path)
      x-readiness: implemented
      x-owner-spec: "spec/02; runcoach-api F003"
      x-story: [sync-completed-from-garmin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary, description: "A .fit file, ≤ 50 MB." }
                resting_capture:
                  type: boolean
                  default: false
                  description: >
                    F004 upload-time Tier-1 declaration — the athlete asserting this
                    file was a resting-HRV capture (config-based profile names cannot
                    reclassify an already-recorded file). Optional; a real bool
                    (true/1/on/yes and negatives coerced), anything else → 422.
      responses:
        "201":
          description: Ingested (or re-derived deterministically)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestResponse" }
        "400":
          description: Not a FIT file / parse failure / a required canonical field could not be determined
          content: { text/plain: { schema: { type: string } } }
        "409":
          description: Duplicate — already ingested (body names the existing session id)
          content: { text/plain: { schema: { type: string } } }
        "413":
          description: Upload exceeds the size/record/beat limit
          content: { text/plain: { schema: { type: string } } }
      # NOTE (as-built): errors are FastAPI-default {detail: "..."} / plain text,
      # not the Error envelope below. Alignment #3 migrates these to Error.

    get:
      tags: [Sessions]
      operationId: listSessions
      summary: List ingested sessions (summaries), newest first
      x-readiness: planned
      x-owner-spec: "spec/02"
      x-story: [view-calendar, view-workout]
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page of session summaries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/SessionSummary" }
  /sessions/{session_id}:
    get:
      tags: [Sessions]
      operationId: getSession
      summary: Full canonical session (header + records + rr_intervals + context)
      description: >
        Raw-over-derived: returns the vendor-neutral canonical streams. The
        quarantined vendor sidecar (VO2max, Training Status, …) is deliberately
        excluded (spec/02 §2.3.6). As-built this returns an untyped dict;
        alignment #1 types it as SessionDetail (shape below is exact to the
        repo's models.py / db.py).
      x-readiness: implemented
      x-owner-spec: "spec/02 (canonical schema)"
      x-story: [view-workout]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Canonical session detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Sessions]
      operationId: deleteSession
      summary: Delete a session and every row that hangs off it
      description: >
        The recovery path out of F004's upgrade window: a capture ingested before
        the athlete declared its profile carries no reading and a re-upload hits
        the UNIQUE (source_device, start_time) 409. Deleting frees the slot so
        "configure, delete, upload again" works. (A destructive control — the UI
        gates it behind explicit confirmation.)
      x-readiness: implemented
      x-owner-spec: "runcoach-api F004 (2026-09-06)"
      x-story: [sync-completed-from-garmin]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------- Workouts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Planned. Backend is single-athlete/no-auth today; alignment #2 adds this."
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string }
      description: Deterministic id = truncated SHA-256 over (source_device, start_time).
...
    Since:
      name: since
      in: query
      required: false
      schema: { type: string }
      description: Opaque cursor from a prior page's next_cursor.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      description: Uniform error envelope (alignment #3). As-built endpoints currently return FastAPI {detail}.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: not_found }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Page:
      type: object
      description: Cursor pagination envelope.
      required: [items]
      properties:
        items: { type: array, items: {} }
        next_cursor:
          type: string
          nullable: true
          description: Pass as `since` to fetch the next page; null when exhausted.
...
    IngestResponse:
      type: object
      required: [session_id, quality_flags]
      properties:
        session_id: { type: string }
        quality_flags:
          type: array
          items: { type: string }
          description: Data-quality flags raised at ingest (never hidden). See spec/02 §2.4.
    SessionDetail:
      type: object
      description: Canonical session as GET /sessions/{id} returns it (quarantine sidecar excluded).
      required: [session_id, start_time, sport, source_vendor, records, rr_intervals]
      properties:
        session_id: { type: string }
        athlete_id: { type: string, nullable: true }
        start_time: { type: string, description: "ISO-8601." }
        sport: { type: string }
        activity_tag: { type: string, nullable: true }
        source_vendor: { type: string }
        source_device: { type: string, nullable: true }
        recording_interval:
          type: string
          nullable: true
          description: Descriptor, not a period — "1hz" / "smart" / "irregular" (spec/02 §2.2.3).
        hr_source: { type: string, nullable: true }
        rr_valid_fraction:
          type: number
          nullable: true
          description: Surviving-beat fraction; null when no RR stream (distinct from 0.0).
        rmssd_precomputed: { type: number, nullable: true, description: "Device-ONLY audit record of resting rMSSD (ms) — the device value on Tier 2, null on Tier 1. Not the engine's read; see resting_rmssd_ms." }
        resting_rmssd_ms:
          type: number
          nullable: true
          exclusiveMinimum: 0
          x-invariant: >
            Scoped, not table-wide. Post-amendment rows (written 2026-09-06+):
            hrv_source_tier != null  ⇒  resting_rmssd_ms != null AND > 0 (both
            tiers write tier and reading together past the non-positive gate).
            Pre-amendment Tier-1 rows may have hrv_source_tier != null WITH
            resting_rmssd_ms == null (the amendment resolved the rule, not the
            rows; no backfill). So a non-null tier does NOT imply a non-null
            reading across the window.
          description: >
            RESOLVED resting rMSSD (ms) the engine (E003) reads on either tier —
            device value on Tier 2, system-computed from beats on Tier 1 (added
            2026-09-06). Nullable, and when present strictly positive
            (exclusiveMinimum 0). Per x-invariant, a non-null hrv_source_tier does
            NOT guarantee a non-null value across the amendment window, so
            consumers guard the read (never assume ln() is safe) and the UI
            surfaces a null reading as provisional/absent, never as zero. This is
            the engine's read; rmssd_precomputed is only the device-value audit.
        hrv_source_tier:
          type: string
          nullable: true
          enum: [chest_strap_raw, health_snapshot, health_api_overnight]
          description: >
            Which tier produced the reading. Non-null does not by itself imply a
            non-null resting_rmssd_ms — see that field's x-invariant.
        rr_source: { type: string, nullable: true, description: "Session-level RR-stream source (spec/02 §2.2.3)." }
        quality_flags: { type: array, items: { type: string } }
        summary: { type: object, nullable: true, additionalProperties: true }
        context: { $ref: "#/components/schemas/SessionContext" }
        records:
          type: array
          items: { $ref: "#/components/schemas/Record" }
        rr_intervals:
          type: array
          items: { $ref: "#/components/schemas/RRInterval" }
    Record:
      type: object
      description: One per-sample canonical record (raw streams).
      required: [t]
      properties:
        t: { type: number, description: "Seconds from session start." }
        lat: { type: number, nullable: true }
        lon: { type: number, nullable: true }
        distance: { type: number, nullable: true }
        speed: { type: number, nullable: true }
        heart_rate: { type: number, nullable: true }
        cadence: { type: number, nullable: true }
        altitude: { type: number, nullable: true }
        power: { type: number, nullable: true }
        power_model: { type: string, nullable: true }
        vertical_oscillation: { type: number, nullable: true }
        ground_contact_time: { type: number, nullable: true }
        gct_balance: { type: number, nullable: true }
        step_length: { type: number, nullable: true }
        temperature: { type: number, nullable: true, description: "Device thermistor, not environmental truth." }
        gps_degraded: { type: boolean, nullable: true }
        sample_quality: { type: array, items: { type: string } }
    RRInterval:
      type: object
      properties:
        seq: { type: integer, nullable: true }
        rr_ms: { type: number, nullable: true }
        rr_source: { type: string, nullable: true, description: "Per-beat tier (spec/02 §2.2.3)." }
        rr_carrier: { type: string, nullable: true }
        is_artefact: { type: boolean, nullable: true }
    SessionContext:
      type: object
      nullable: true
      description: Environmental + provenance metadata (spec/02 §2.2.4). env_* require an external weather source, null on the FIT-upload route.
      properties:
        ingested_at: { type: string, nullable: true }
        provenance: { type: object, additionalProperties: true }
        env_temperature_c: { type: number, nullable: true }
        env_humidity_pct: { type: number, nullable: true }
        env_wind_ms: { type: number, nullable: true }
        env_wind_dir: { type: string, nullable: true }
        subjective: { type: object, nullable: true, additionalProperties: true }
    SessionSummary:
      type: object
      description: List-view projection of a session (planned shape).
      properties:
        session_id: { type: string }
        start_time: { type: string }
        sport: { type: string }
        activity_tag: { type: string, nullable: true }
        distance_m: { type: number, nullable: true }
        duration_s: { type: number, nullable: true }
        avg_pace_s_per_km: { type: number, nullable: true }
        quality_flags: { type: array, items: { type: string } }
        workout_id: { type: string, nullable: true, description: "Linked planned workout, if this activity executed one." }

    # ---- Athlete / onboarding (planned — field-level per spec/04 §4.4) ---
...